moved functions to YoutubeAPI module
