+++ lib/video_info/providers/youtube.rb
@@ -1,7 +1,15 @@
 require 'iso8601'
+require_relative 'youtube_api'
+
 class VideoInfo
   module Providers
     class Youtube < Provider
+      def initialize(url, options = {})
+        extend YoutubeAPI
+
+        super(url, options)
+      end
+
       def self.usable?(url)
         url =~ /(youtube\.com\/(?!playlist|embed\/videoseries).*)|(youtu\.be)/
       end
@@ -10,38 +18,12 @@ class VideoInfo
         'YouTube'
       end
 
-      def api_key
-        VideoInfo.provider_api_keys[:youtube]
-      end
-
-      def title
-        _video_snippet['title']
-      end
-
-      def description
-        _video_snippet['description']
-      end
-
-      def keywords
-        _video_snippet['tags']
-      end
-
       %w[width height].each do |method|
         define_method(method) { nil }
       end
 
-      def duration
-        video_duration = _video_content_details['duration'] || 0
-        ISO8601::Duration.new(video_duration).to_seconds.to_i
-      end
-
       def embed_url
         "//www.youtube.com/embed/#{video_id}"
-      end
-
-      def date
-        return unless published_at = _video_snippet['publishedAt']
-        Time.parse(published_at, Time.now.utc)
       end
 
       def thumbnail_small
@@ -59,30 +41,10 @@ class VideoInfo
         _video_snippet['thumbnails']['high']['url']
       end
 
-      def view_count
-        _video_statistics['viewCount'].to_i rescue 0
-      end
-
       private
-
-      def available?
-        data['items'].size > 0 rescue false
-      end
 
       def _url_regex
         /(?:youtube(?:-nocookie)?\.com\/(?:[^\/]+\/.+\/|(?:v|e(?:mbed)?)\/|.*[?&]v=)|youtu\.be\/)([^"&?\/ ]{11})/i
-      end
-
-      def _api_base
-        'www.googleapis.com'
-      end
-
-      def _api_path
-        "/youtube/v3/videos?id=#{video_id}&part=snippet,statistics,contentDetails&fields=items(id,snippet,statistics,contentDetails)&key=#{api_key}"
-      end
-
-      def _api_url
-        "https://#{_api_base}#{_api_path}"
       end
 
       def _default_iframe_attributes
@@ -91,25 +53,6 @@ class VideoInfo
 
       def _default_url_attributes
         {}
-      end
-
-      def _video_snippet
-        return {} unless available?
-        data['items'][0]['snippet']
-      end
-
-      def _video_content_details
-        return {} unless available?
-        data['items'][0]['contentDetails']
-      end
-
-      def _video_statistics
-        return {} unless available?
-        data['items'][0]['statistics']
-      end
-
-      def _video_thumbnail(id)
-        _video_entry['media$group']['media$thumbnail'][id]['url']
       end
     end
   end
